Added "Odin's possesivness to basic red sauce titles. Also added a css file for the index.html
Centered the elements in the index.html in a css file called "indexstyle.css". Linkned the css file externally, potential use for the other html files for the food recipes maybe used in the future. Also added the images of the food below the links for that recipe in the index.html, adjusted their size in the css file to take up less space and so image size was uniform. Made it so the links do no show an underline unless hovering over the link.